chore: Update imports for resolver move
