+++ benches/benchsuite/benches/resolve.rs
@@ -1,10 +1,10 @@
 use benchsuite::fixtures;
 use cargo::GlobalContext;
 use cargo::core::compiler::{CompileKind, RustcTargetData};
-use cargo::core::resolver::features::{FeatureOpts, FeatureResolver};
-use cargo::core::resolver::{CliFeatures, ForceAllTargets, HasDevUnits, ResolveBehavior};
 use cargo::core::{PackageIdSpec, Workspace};
 use cargo::ops::WorkspaceResolve;
+use cargo::resolver::features::{FeatureOpts, FeatureResolver};
+use cargo::resolver::{CliFeatures, ForceAllTargets, HasDevUnits, ResolveBehavior};
 use criterion::{Criterion, criterion_group, criterion_main};
 use std::path::Path;
