+++ src/core/mod.rs
@@ -6,14 +6,12 @@ pub use self::package::{Package, PackageSet};
 pub use self::package_id::PackageId;
 pub use self::package_id_spec::PackageIdSpecQuery;
 pub use self::registry::Registry;
-pub use self::resolver::{Resolve, ResolveVersion};
 pub use self::source_id::SourceId;
 pub use self::summary::{FeatureMap, FeatureValue, Summary};
 pub use self::workspace::{
     MaybePackage, Workspace, WorkspaceConfig, WorkspaceRootConfig, find_workspace_root,
     find_workspace_root_with_membership_check, resolve_relative_path,
 };
-pub use crate::resolver;
 pub use cargo_util_schemas::core::{GitReference, PackageIdSpec, SourceKind};
 
 pub mod compiler;
