+++ src/core/package.rs
@@ -21,13 +21,13 @@ use tracing::debug;
 
 use crate::core::compiler::{CompileKind, RustcTargetData};
 use crate::core::dependency::DepKind;
-use crate::core::resolver::features::ForceAllTargets;
-use crate::core::resolver::{HasDevUnits, Resolve};
 use crate::core::{
     CliUnstable, Dependency, Features, Manifest, PackageId, PackageIdSpec, SerializedDependency,
     SourceId, Target,
 };
 use crate::core::{Summary, Workspace};
+use crate::resolver::features::ForceAllTargets;
+use crate::resolver::{HasDevUnits, Resolve};
 use crate::sources::source::{MaybePackage, SourceMap};
 use crate::util::HumanBytes;
 use crate::util::cache_lock::{CacheLock, CacheLockMode};
