+++ src/core/profiles.rs
@@ -24,8 +24,9 @@
 use crate::core::Feature;
 use crate::core::compiler::{CompileKind, CompileTarget, Unit};
 use crate::core::dependency::Artifact;
-use crate::core::resolver::features::FeaturesFor;
-use crate::core::{PackageId, PackageIdSpec, PackageIdSpecQuery, Resolve, Target, Workspace};
+use crate::core::{PackageId, PackageIdSpec, PackageIdSpecQuery, Target, Workspace};
+use crate::resolver::Resolve;
+use crate::resolver::features::FeaturesFor;
 use crate::util::data_structures::{HashMap, HashSet};
 use crate::util::interning::InternedString;
 use crate::util::toml::validate_profile;
