+++ src/ops/cargo_add/mod.rs
@@ -30,8 +30,8 @@ use crate::core::Summary;
 use crate::core::Workspace;
 use crate::core::dependency::DepKind;
 use crate::core::registry::PackageRegistry;
-use crate::core::resolver::PublishAgePolicy;
 use crate::ops::resolve_ws;
+use crate::resolver::PublishAgePolicy;
 use crate::sources::IndexSummary;
 use crate::sources::source::QueryKind;
 use crate::util::OptVersionReq;
