+++ src/ops/cargo_fetch.rs
@@ -2,8 +2,9 @@ use crate::core::compiler::BuildConfig;
 use crate::core::compiler::RustcTargetData;
 use crate::core::compiler::UserIntent;
 use crate::core::compiler::standard_lib;
-use crate::core::{PackageSet, Resolve, Workspace};
+use crate::core::{PackageSet, Workspace};
 use crate::ops;
+use crate::resolver::Resolve;
 use crate::util::CargoResult;
 use crate::util::GlobalContext;
 use crate::util::context::JobsConfig;
