+++ src/ops/cargo_package/mod.rs
@@ -11,11 +11,12 @@ use crate::core::PackageIdSpecQuery;
 use crate::core::Workspace;
 use crate::core::dependency::DepKind;
 use crate::core::manifest::Target;
-use crate::core::resolver::CliFeatures;
-use crate::core::resolver::HasDevUnits;
-use crate::core::{Package, PackageId, PackageSet, Resolve, SourceId};
+use crate::core::{Package, PackageId, PackageSet, SourceId};
 use crate::ops::lockfile::LOCKFILE_NAME;
 use crate::ops::registry::{RegistryOrIndex, infer_registry};
+use crate::resolver::CliFeatures;
+use crate::resolver::HasDevUnits;
+use crate::resolver::Resolve;
 use crate::sources::path::PathEntry;
 use crate::sources::source::QueryKind;
 use crate::sources::{CRATES_IO_REGISTRY, PathSource};
