+++ src/ops/registry/cargo_publish.rs
@@ -31,13 +31,13 @@ use crate::core::SourceId;
 use crate::core::Workspace;
 use crate::core::dependency::DepKind;
 use crate::core::manifest::ManifestMetadata;
-use crate::core::resolver::CliFeatures;
 use crate::ops;
 use crate::ops::PackageOpts;
 use crate::ops::Packages;
 use crate::ops::RegistryOrIndex;
 use crate::ops::registry::RegistryClient;
 use crate::ops::registry::RegistrySourceIds;
+use crate::resolver::CliFeatures;
 use crate::sources::CRATES_IO_REGISTRY;
 use crate::sources::RegistrySource;
 use crate::sources::SourceConfigMap;
